requirements.txt
django==3.1.7
Pillow==8.1.2
django-debug-toolbar==3.2